for some reason dlr removed the adding of junit to the classpath, however, the <available> task fails to find junit because it isn't in the classpath. we still have the problem of not being able to run junit unless it is in the system classpath though because the <junit> task doens't find junit and produces this error:...:-(
/Users/jon/checkout/xml-rpc/build.xml:247: Could not create task of
type: junit. Common solutions are to use taskdef to declare your task,
or, if this is an optional task, to put the optional.jar in the lib
directory of your ant installation (ANT_HOME).

I bet the solution is to put junit.jar into $ANT_HOME/lib and forget
about it...

-jon