Applied patch by Kristian Meier <kristianm@jentro.com>.  The following two descriptions were supplied in different messages:
"we are going to use the webserver in our new release. for this we need
SSL. there is a quite convinient tool for setting up the neccassary
ssl-system-properties, the SecurityTools. but unfortunatly there is no
setter-method for the TrustStorePassword and the getter uses the
DEFAULT_KEY_PASSWORD if not set."

"there is no setter for the trustStorePassword, so I included it along
with javadocs.

the second one is just a typo, where the keyStorePassword is checked on
null, this check must be on the trustStorePassword, since we are in the
getTrustStorePassword-method."