Worker.executeInternal(): Clarified debug message, broke long line, and re-add block removed during Adam's refactoring.
Invoker.execute(): If the method name contains dots, use only the last
dotted element as the Java method name (the rest is the XML-RPC method
name, which is also used to look up which handler class to use).  This
used to occur in Worker.executeInternal(), but was removed by Adam's
refactoring.  Things are more clear and somewhat better documented
now, so it's all for the best.